fix: 修復圖書館新版端點造成的 API 錯誤
為了應對可能發生的改變，get_library_opening_hours 改為單純回應 dict。
